Create useReadPosts hook
